# lab10/src/main/resources/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/rulers":{"post":{"tags":["ruler-controller"],"operationId":"addRuler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RulerDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ruler"}}}}}}},"/api/passports":{"post":{"tags":["passport-controller"],"operationId":"addPassport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PassportDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Passport"}}}}}}},"/api/countries":{"post":{"tags":["country-controller"],"operationId":"addCountry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/citizens":{"post":{"tags":["citizen-controller"],"operationId":"addCitizen","parameters":[{"name":"citizen","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CitizenDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Citizen"}}}}}}},"/api/cities":{"post":{"tags":["city-controller"],"operationId":"addCity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/City"}}}}}}},"/api/rulers/{id}":{"get":{"tags":["ruler-controller"],"operationId":"getRulerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ruler"}}}}}},"delete":{"tags":["ruler-controller"],"operationId":"deleteRuler","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ruler"}}}}}}},"/api/passports/{id}":{"get":{"tags":["passport-controller"],"operationId":"getPassportById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Passport"}}}}}},"delete":{"tags":["passport-controller"],"operationId":"deletePassport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Passport"}}}}}}},"/api/countries/{id}":{"get":{"tags":["country-controller"],"operationId":"getCountryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}},"delete":{"tags":["country-controller"],"operationId":"deleteCountry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/countries/findByNameOrFormationYear":{"get":{"tags":["country-controller"],"operationId":"getByNameOrFormationYear","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"formationYear","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/countries/findByNameAndFormationYear":{"get":{"tags":["country-controller"],"operationId":"getByNameAndFormationYear","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"formationYear","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Country"}}}}}}},"/api/countries/all":{"get":{"tags":["country-controller"],"operationId":"getAllCountries","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/citizens/{id}":{"get":{"tags":["citizen-controller"],"operationId":"getCitizenById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Citizen"}}}}}},"delete":{"tags":["citizen-controller"],"operationId":"deleteCitizen","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Citizen"}}}}}}},"/api/cities/{id}":{"get":{"tags":["city-controller"],"operationId":"getCityById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/City"}}}}}},"delete":{"tags":["city-controller"],"operationId":"deleteCity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/City"}}}}}}}},"components":{"schemas":{"RulerDTO":{"type":"object","properties":{"name":{"type":"string"},"surname":{"type":"string"},"officeStartDate":{"type":"string","format":"date-time"}}},"Citizen":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"lastName":{"type":"string"},"countries":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Country"}},"passport":{"$ref":"#/components/schemas/Passport"}}},"City":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"$ref":"#/components/schemas/Country"}}},"Country":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"gdp":{"type":"integer","format":"int64"},"formationYear":{"type":"object","properties":{"value":{"type":"integer","format":"int32"},"leap":{"type":"boolean"}}},"ruler":{"$ref":"#/components/schemas/Ruler"},"cities":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/City"}},"citizens":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Citizen"}},"inEurope":{"type":"boolean"}}},"Passport":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"serialNumber":{"type":"string"},"citizen":{"$ref":"#/components/schemas/Citizen"}}},"Ruler":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surname":{"type":"string"},"officeStartDate":{"type":"string","format":"date-time"},"country":{"$ref":"#/components/schemas/Country"}}},"PassportDTO":{"type":"object","properties":{"serialNumber":{"type":"string"}}},"CountryDTO":{"type":"object","properties":{"name":{"type":"string"},"gdp":{"type":"integer","format":"int64"},"isInEurope":{"type":"boolean"},"formationYear":{"type":"object","properties":{"value":{"type":"integer","format":"int32"},"leap":{"type":"boolean"}}}}},"CitizenDTO":{"type":"object","properties":{"name":{"type":"string"},"lastName":{"type":"string"}}},"CityDTO":{"type":"object","properties":{"name":{"type":"string"}}}}}}